refactor: 달력 fallback 컴포넌트에서 카테고리 미선택 분리
