+++ frontend/src/components/Calendar/Calendar.styles.ts
@@ -2,23 +2,6 @@ import { css, Theme } from '@emotion/react';
 
 import { DAYS } from '@/constants/date';
 
-const hintStyle = ({ colors }: Theme) => css`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%);
-  z-index: 10;
-
-  padding: 4rem 6rem;
-  border-radius: 7px;
-
-  background: ${colors.ORANGE_500};
-
-  font-size: 4rem;
-  font-weight: 500;
-  color: ${colors.WHITE};
-`;
-
 const calendarHeaderStyle = ({ colors, flex }: Theme) => css`
   ${flex.row}
 
@@ -138,7 +121,6 @@ export {
   calendarHeaderStyle,
   dayGridStyle,
   dayStyle,
-  hintStyle,
   monthPickerStyle,
   navButtonStyle,
   navButtonTitleStyle,
