+++ frontend/src/pages/CategoryPage/CategoryPage.styles.ts
@@ -19,4 +19,21 @@ const calendarStyle = ({ mq }: Theme) => css`
   }
 `;
 
-export { calendarStyle, categoryPageStyle };
+const hintStyle = ({ colors }: Theme) => css`
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%);
+  z-index: 10;
+
+  padding: 4rem 6rem;
+  border-radius: 7px;
+
+  background: ${colors.ORANGE_500};
+
+  font-size: 4rem;
+  font-weight: 500;
+  color: ${colors.WHITE};
+`;
+
+export { calendarStyle, categoryPageStyle, hintStyle };
