+++ frontend/src/pages/CategoryPage/CategoryPage.tsx
@@ -12,7 +12,7 @@ import CategoryControl from '@/components/CategoryControl/CategoryControl';
 
 import { PAGE_LAYOUT } from '@/constants/style';
 
-import { calendarStyle, categoryPageStyle } from './CategoryPage.styles';
+import { calendarStyle, categoryPageStyle, hintStyle } from './CategoryPage.styles';
 
 function CategoryPage() {
   const [category, setCategory] = useState<Pick<CategoryType, 'id' | 'name'>>({ id: 0, name: '' });
@@ -26,15 +26,29 @@ function CategoryPage() {
     endDateTime,
   });
 
+  if (category.id === 0) {
+    return (
+      <PageLayout type={PAGE_LAYOUT.SIDEBAR}>
+        <div css={categoryPageStyle}>
+          <CategoryControl setCategory={setCategory} />
+          <div css={calendarStyle}>
+            <div css={hintStyle}>클릭한 카테고리의 일정을 확인할 수 있어요</div>
+            <CalendarFallback calendarController={calendarController} isLoading={false} readonly />
+          </div>
+        </div>
+      </PageLayout>
+    );
+  }
+
   return (
     <PageLayout type={PAGE_LAYOUT.SIDEBAR}>
       <div css={categoryPageStyle}>
         <CategoryControl setCategory={setCategory} />
         <div css={calendarStyle}>
-          {(isLoading || data === undefined) && (
+          {isLoading && (
             <CalendarFallback
               calendarController={calendarController}
-              category={category}
+              categoryName={category.name}
               readonly
             />
           )}
@@ -42,7 +56,7 @@ function CategoryPage() {
             <Calendar
               calendarController={calendarController}
               scheduleResponse={data}
-              category={category}
+              categoryName={category.name}
               readonly
             />
           )}
